+++ CHANGELOG.md
@@ -61,6 +61,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - undocumented prefix feature in conversation
 - `citizens_npcs_by_name` config option, which is now part of the id
 - start and stop sound in the configuration for conversations were removed in favor of the notification system that now also has the two new build in categories `conversation_start` and `conversation_end`
+- `menu reload` command which did not work
 ### Fixed
 - Reloading plugin did not reload Menu config
 - potions generated not extended/upgraded since 2.2.1